requirements.txt
aniso8601==9.0.1
async-timeout==4.0.3
bidict==0.23.1
blinker==1.7.0
cachelib==0.12.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.10
Flask-Session==0.6.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
mysqlclient==2.2.4
psutil==5.9.8
PyJWT==2.8.0
python-engineio==4.9.0
python-socketio==5.11.1
pytz==2024.1
redis==5.0.3
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
Werkzeug==3.0.1
wsproto==1.2.0
